cephadm: move executable discovery funcs to exe_utils.py
Pair-programmed-with: Adam King <adking@redhat.com>
