+++ src/cephadm/cephadm.py
@@ -98,6 +98,7 @@ from cephadmlib.exceptions import (
     TimeoutExpired,
     UnauthorizedRegistryError,
 )
+from cephadmlib.exe_utils import find_executable, find_program
 
 FuncT = TypeVar('FuncT', bound=Callable)
 
@@ -2630,51 +2631,6 @@ def recursive_chown(path: str, uid: int, gid: int) -> None:
         os.chown(dirpath, uid, gid)
         for filename in filenames:
             os.chown(os.path.join(dirpath, filename), uid, gid)
-
-
-# copied from distutils
-def find_executable(executable: str, path: Optional[str] = None) -> Optional[str]:
-    """Tries to find 'executable' in the directories listed in 'path'.
-    A string listing directories separated by 'os.pathsep'; defaults to
-    os.environ['PATH'].  Returns the complete filename or None if not found.
-    """
-    _, ext = os.path.splitext(executable)
-    if (sys.platform == 'win32') and (ext != '.exe'):
-        executable = executable + '.exe'  # pragma: no cover
-
-    if os.path.isfile(executable):
-        return executable
-
-    if path is None:
-        path = os.environ.get('PATH', None)
-        if path is None:
-            try:
-                path = os.confstr('CS_PATH')
-            except (AttributeError, ValueError):
-                # os.confstr() or CS_PATH is not available
-                path = os.defpath
-        # bpo-35755: Don't use os.defpath if the PATH environment variable is
-        # set to an empty string
-
-    # PATH='' doesn't match, whereas PATH=':' looks in the current directory
-    if not path:
-        return None
-
-    paths = path.split(os.pathsep)
-    for p in paths:
-        f = os.path.join(p, executable)
-        if os.path.isfile(f):
-            # the file exists, we have a shot at spawn working
-            return f
-    return None
-
-
-def find_program(filename):
-    # type: (str) -> str
-    name = find_executable(filename)
-    if name is None:
-        raise ValueError('%s not found' % filename)
-    return name
 
 
 def find_container_engine(ctx: CephadmContext) -> Optional[ContainerEngine]:
